test: 更新react-native-translucent-modal三方库的测试用例和基本使用场景Demo
